+++ src/main/java/barch/tzm/Config/TheZombieModConfig.java
@@ -6,7 +6,7 @@ public class TheZombieModConfig {
     @Comment("The amount of health zombies have. DEFAULT: [5]")
     public int zombieHealth = 5;
 
-    @Comment("Zombies replace normal enemies. If true, they will spawn practicaly replacing all normal overworld enemies. If false, they will spawn ofter butonly like most nermal mobs. DEFAULT [true]")
+    @Comment("Zombies replace normal enemies. If true, they will spawn practically replacing all normal overworld enemies. If false, they will spawn often but only like most normal mobs. DEFAULT [true]")
     public Boolean zombiesTakeOver = true;
 
 }